alter rank field to floatin point number

+++ backend_core/players/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Player(models.Model):
     nick = models.CharField(max_length=30)
-    rank = models.PositiveSmallIntegerField()
+    rank = models.FloatField()
     is_bot = models.BooleanField(default=False)
 
     def __str__(self):
